+++ CHANGELOG.md
@@ -385,3 +385,5 @@
 - UI Anpassungen: Gießen-Button wurde wieder etwas verkleinert und weiter nach oben verschoben, um nicht mit der Tab-Bar zu kollidieren.
 
 - UI Anpassungen: Gießen-Button wurde wieder an seine ursprüngliche Position verschoben und das Wassertropfen-Icon im Button leicht nach oben gerückt.
+
+- UI Anpassungen: Gießen-Button wurde wieder etwas verkleinert.